Add conftest.py so pytest can find Python modules
When I moved the tests to the tests/ directory, pytest started throwing
a `ModuleNotFoundError`. Pytest could find the tests, but the tests
couldn't find the modules they were importing from the root directory.
A Stack Overflow solution https://stackoverflow.com/questions/10253826
is to simply create an empty _conftest.py_ file in the root directory.

This seems strange to me.