Switch from Pipenv to Poetry
This project now uses Poetry for dependency management.
https://python-poetry.org/

Pipenv was previously used to manage the development virtual environment
for this project. https://pipenv.readthedocs.io/en/latest/

The future of Pipenv is unclear. It went all of 2019 without a major
release, and has many bugs. Jacob Kaplan-Moss (Django co-creator) has
commented on how "the lead of Pipenv was someone with a history of not
treating his collaborators well," and on the "bugs and rapid API
changes" of Pipenv.

Jacob Kaplan-Moss: My Python development environment, 2020 edition
https://jacobian.org/2019/nov/11/python-environment-2020/