Calculate version number based on pyproject.toml
During `poetry install`, Poetry installs the project as if it were a
package itself. You can then pull the version from the package metadata.
It's a little tricky to get `importlib_metadata` to work with Docker.
You have to remove `--no-root` from the _Dockerfile_ and add
`importlib-metadata` to prod dependencies, and you still might get
errors after that.

https://github.com/python-poetry/poetry/issues/144
https://github.com/python-poetry/poetry/issues/1036